examples/classification/requirements.txt
# 基础依赖
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.20.0
matplotlib>=3.4.0
pillow>=8.0.0
opencv-python>=4.5.0
openmim>=0.3.7

# MMPretrain及其依赖
mmpretrain>=1.0.0
mmengine>=0.7.0
mmcv>=2.0.0

# 高级示例依赖 (可选)
# 如果安装失败，可以注释掉下面这行，高级示例中的CAM可视化功能将不可用
# pytorch-grad-cam>=1.3.7
grad-cam>=1.4.0  # 替代方案

# 如果上面的grad-cam也安装失败，可以尝试手动安装：
# pip install git+https://github.com/jacobgil/pytorch-grad-cam.git 

# python -c "import sys; print(sys.executable)" 查询python安装目录